requirements.txt
httpx
selectolax
pyaml
aiofiles
